# Assignment 2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GitHub Issues Gateway",
    "version": "1.0.0"
  },
  "paths": {
    "/issues/{number}": {
      "get": {
        "summary": "Get Issue Endpoint",
        "description": "Return a single issue by issue number.",
        "operationId": "get_issue",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Issue returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                },
                "example": {
                  "number": 4,
                  "html_url": "https://github.com/ChanghyunKim-sjsu/CMPE-272-issues-test/issues/4",
                  "state": "open",
                  "title": "Example issue",
                  "body": "Created through the CMPE 272 Issues Gateway.",
                  "labels": [
                    "bug"
                  ],
                  "created_at": "2026-09-05T20:15:00Z",
                  "updated_at": "2026-09-05T20:15:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Invalid issue number",
            "content": {
              "application/json": {
                "example": {
                  "error": "validation_error",
                  "message": "Invalid request",
                  "status_code": 400
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "GitHub authentication failed",
            "content": {
              "application/json": {
                "example": {
                  "error": "github_api_error",
                  "message": "Bad credentials",
                  "status_code": 401
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Issue not found",
            "content": {
              "application/json": {
                "example": {
                  "error": "github_api_error",
                  "message": "Issue not found",
                  "status_code": 404
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "GitHub API rate limit exceeded",
            "content": {
              "application/json": {
                "example": {
                  "error": "github_api_error",
                  "message": "GitHub API rate limit exceeded",
                  "status_code": 429
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Issue Endpoint",
        "description": "Update an existing issue.",
        "operationId": "update_issue",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Issue updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                },
                "example": {
                  "number": 4,
                  "html_url": "https://github.com/ChanghyunKim-sjsu/CMPE-272-issues-test/issues/4",
                  "state": "open",
                  "title": "Example issue",
                  "body": "Created through the CMPE 272 Issues Gateway.",
                  "labels": [
                    "bug"
                  ],
                  "created_at": "2026-09-05T20:15:00Z",
                  "updated_at": "2026-09-05T20:15:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Invalid issue update payload",
            "content": {
              "application/json": {
                "example": {
                  "error": "validation_error",
                  "message": "Invalid request",
                  "status_code": 400
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "GitHub authentication failed",
            "content": {
              "application/json": {
                "example": {
                  "error": "github_api_error",
                  "message": "Bad credentials",
                  "status_code": 401
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Issue not found",
            "content": {
              "application/json": {
                "example": {
                  "error": "github_api_error",
                  "message": "Issue not found",
                  "status_code": 404
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "GitHub API rate limit exceeded",
            "content": {
              "application/json": {
                "example": {
                  "error": "github_api_error",
                  "message": "GitHub API rate limit exceeded",
                  "status_code": 429
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/issues": {
      "get": {
        "summary": "List Issues Endpoint",
        "description": "Return issues from the configured GitHub repository.",
        "operationId": "list_issues",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "open",
                "closed",
                "all"
              ],
              "type": "string",
              "default": "open",
              "title": "State"
            }
          },
          {
            "name": "labels",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Labels"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Issues returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IssueResponse"
                  },
                  "title": "Response List Issues"
                },
                "example": [
                  {
                    "number": 4,
                    "html_url": "https://github.com/ChanghyunKim-sjsu/CMPE-272-issues-test/issues/4",
                    "state": "open",
                    "title": "Example issue",
                    "body": "Created through the CMPE 272 Issues Gateway.",
                    "labels": [
                      "bug"
                    ],
                    "created_at": "2026-09-05T20:15:00Z",
                    "updated_at": "2026-09-05T20:15:00Z"
                  }
                ]
              }
            },
            "headers": {
              "Link": {
                "description": "GitHub pagination links",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "example": {
                  "error": "validation_error",
                  "message": "Invalid request",
                  "status_code": 400
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "GitHub authentication failed",
            "content": {
              "application/json": {
                "example": {
                  "error": "github_api_error",
                  "message": "Bad credentials",
                  "status_code": 401
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "GitHub API rate limit exceeded",
            "content": {
              "application/json": {
                "example": {
                  "error": "github_api_error",
                  "message": "GitHub API rate limit exceeded",
                  "status_code": 429
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Issue Endpoint",
        "description": "Create a new issue in the configured GitHub repository.",
        "operationId": "create_issue",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Issue created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                },
                "example": {
                  "number": 4,
                  "html_url": "https://github.com/ChanghyunKim-sjsu/CMPE-272-issues-test/issues/4",
                  "state": "open",
                  "title": "Example issue",
                  "body": "Created through the CMPE 272 Issues Gateway.",
                  "labels": [
                    "bug"
                  ],
                  "created_at": "2026-09-05T20:15:00Z",
                  "updated_at": "2026-09-05T20:15:00Z"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "Location of the newly created issue",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid issue payload",
            "content": {
              "application/json": {
                "example": {
                  "error": "validation_error",
                  "message": "Invalid request",
                  "status_code": 400
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "GitHub authentication failed",
            "content": {
              "application/json": {
                "example": {
                  "error": "github_api_error",
                  "message": "Bad credentials",
                  "status_code": 401
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "GitHub API rate limit exceeded",
            "content": {
              "application/json": {
                "example": {
                  "error": "github_api_error",
                  "message": "GitHub API rate limit exceeded",
                  "status_code": 429
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message",
          "status_code"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response returned by the API.",
        "examples": [
          {
            "error": "github_api_error",
            "message": "Issue not found",
            "status_code": 404
          }
        ]
      },
      "IssueCreate": {
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Labels"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "IssueCreate",
        "description": "Request model for creating a GitHub issue.",
        "examples": [
          {
            "body": "Created through the CMPE 272 Issues Gateway.",
            "labels": [
              "bug",
              "assignment"
            ],
            "title": "Example issue"
          }
        ]
      },
      "IssueResponse": {
        "properties": {
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "html_url": {
            "type": "string",
            "title": "Html Url"
          },
          "state": {
            "type": "string",
            "enum": [
              "open",
              "closed"
            ],
            "title": "State"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Labels"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "number",
          "html_url",
          "state",
          "title",
          "body",
          "labels",
          "created_at",
          "updated_at"
        ],
        "title": "IssueResponse",
        "description": "Response model returned by the Issues Gateway.",
        "examples": [
          {
            "body": "Created through the CMPE 272 Issues Gateway.",
            "created_at": "2026-09-05T20:15:00Z",
            "html_url": "https://github.com/ChanghyunKim-sjsu/CMPE-272-issues-test/issues/4",
            "labels": [
              "bug"
            ],
            "number": 4,
            "state": "open",
            "title": "Example issue",
            "updated_at": "2026-09-05T20:15:00Z"
          }
        ]
      },
      "IssueUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "state": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "open",
                  "closed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          }
        },
        "type": "object",
        "title": "IssueUpdate",
        "description": "Request model for updating a GitHub issue.",
        "examples": [
          {
            "body": "The issue body has been updated.",
            "state": "open",
            "title": "Updated issue title"
          }
        ]
      },
...
    }
  }
}